refactor mouse module into separate files
